chore: Update Nginx configuration to serve a basic webpage
This commit updates the Nginx configuration to serve a basic webpage that returns a "Hello World!" message. It also includes changes to allow Nginx through the firewall and give the user ownership of the website files for easy editing. Additionally, the default index file is backed up and a new index file is created. Finally, the Nginx service is restarted to apply the changes.